+++ packages/Webkul/Core/src/Providers/CoreServiceProvider.php
@@ -59,7 +59,7 @@ class CoreServiceProvider extends ServiceProvider
         });
 
         $this->app->register(EventServiceProvider::class);
-        $this->app->register(DynamicSmtpServiceProvider::class);
+        $this->app->register(DynamicMailServiceProvider::class);
     }
 
     /**